+++ src/signatures.rs
@@ -461,6 +461,9 @@ signatures! {
     value = b"\x7E\x10\x04", b"\x30\x4D\x49\x45" offset = 4
     value = b"\x7E\x18\x04", b"\x30\x4D\x49\x45" offset = 4
 
+    format = Model3dAscii
+    value = b"3dmodel"
+
     format = Mp3Url
     value = b"#EXTM3U"
 
@@ -696,6 +699,9 @@ signatures! {
 
     format = MicrosoftDirectDrawSurface
     value = b"DDS "
+
+    format = Model3dBinary
+    value = b"3DMO"
 
     format = MonkeysAudio
     value = b"MAC "