+++ tests/text.rs
@@ -37,6 +37,12 @@ fn test_lua_script() {
 }
 
 #[test]
+fn test_model_3d_ascii() {
+    let format = FileFormat::from_file("fixtures/text/sample.a3d").unwrap();
+    assert_eq!(format, FileFormat::Model3dAscii);
+}
+
+#[test]
 fn test_ms_dos_batch() {
     let format = FileFormat::from_file("fixtures/text/sample.bat").unwrap();
     assert_eq!(format, FileFormat::MsDosBatch);